# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Authn Platform API
  description: Enterprise Identity Engine REST API specification for authentication, session management, organization multi-tenancy,
    RBAC, SAML, OAuth2/OIDC, Webhooks, API key administration, and account recovery.
  version: 2.0.0
  contact:
    name: Authn Platform Engineering
    url: https://github.com/hanan-bhatti/authn-2.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: http://localhost:8080
  description: Local Development Engine
- url: https://api.staging.authn.dev
  description: Staging Environment
- url: https://api.authn.dev
  description: Production Environment
paths:
...
  /v1/client/auth/2fa/webauthn/register/begin:
    post:
      summary: Begin WebAuthn Passkey Registration
      description: Generates WebAuthn publicKey credential creation options for registering a new passkey.
      operationId: beginWebAuthnRegistration
      tags:
      - WebAuthn Passkeys
      security:
      - PublishableKey: []
        BearerAuth: []
      responses:
        '200':
          description: WebAuthn registration options generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeginWebAuthnRegistrationResponse'
        '400':
          description: Validation failed or unable to begin registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/webauthn/register/finish:
    post:
      summary: Finish WebAuthn Passkey Registration
      description: Verifies authenticator attestation response and saves the new passkey credential.
      operationId: finishWebAuthnRegistration
      tags:
      - WebAuthn Passkeys
      security:
      - PublishableKey: []
        BearerAuth: []
      parameters:
      - in: query
        name: session_id
        required: false
        schema:
          type: string
        description: Ceremony session ID (can alternatively be passed in request body)
      - in: query
        name: name
        required: false
        schema:
          type: string
        description: Optional human-readable passkey name label (can alternatively be passed in request body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebAuthnRegisterFinishRequest'
      responses:
        '200':
          description: Passkey successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmTOTPResponse'
        '400':
          description: Missing session_id or invalid attestation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/webauthn/login/begin:
    post:
      summary: Begin WebAuthn Passkey Login
      description: Generates WebAuthn assertion options to initiate passkey login for a user holding an mfa_token.
      operationId: beginWebAuthnLogin
      tags:
      - WebAuthn Passkeys
      security:
      - PublishableKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebAuthnLoginBeginRequest'
      responses:
        '200':
          description: WebAuthn assertion options generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeginWebAuthnLoginResponse'
        '400':
          description: Missing mfa_token or validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/webauthn/login/finish:
    post:
      summary: Finish WebAuthn Passkey Login
      description: Verifies authenticator assertion response and issues session tokens.
      operationId: finishWebAuthnLogin
      tags:
      - WebAuthn Passkeys
      security:
      - PublishableKey: []
      parameters:
      - in: header
        name: X-Authn-Client-Type
        required: false
        schema:
          type: string
          enum:
          - web
          - native
          - mobile
          default: web
        description: Client platform type. Controls whether refresh token is returned in response body or HttpOnly cookie.
      - in: query
        name: mfa_token
        required: false
        schema:
          type: string
        description: MFA token (can alternatively be passed in request body)
      - in: query
        name: session_id
        required: false
        schema:
          type: string
        description: Assertion ceremony session ID (can alternatively be passed in request body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebAuthnLoginFinishRequest'
      responses:
        '200':
          description: Passkey authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Missing mfa_token/session_id or invalid assertion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/webauthn/credentials:
    get:
      summary: List WebAuthn Passkeys
      description: Returns a list of all registered WebAuthn passkeys for the authenticated user.
      operationId: listWebAuthnPasskeys
      tags:
      - WebAuthn Passkeys
      security:
      - PublishableKey: []
        BearerAuth: []
      responses:
        '200':
          description: List of registered passkeys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPasskeysResponse'
        '401':
          description: Unauthorized session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/webauthn/credentials/{id}:
    delete:
      summary: Delete WebAuthn Passkey
      description: Removes a registered WebAuthn passkey credential. Requires password step-up if deleting the last remaining
        2FA method.
      operationId: deleteWebAuthnPasskey
      tags:
      - WebAuthn Passkeys
      security:
      - PublishableKey: []
        BearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: Internal 2FA method row ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasskeyDeleteRequest'
      responses:
        '200':
          description: Passkey successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletePasskeyResponse'
        '400':
          description: Missing passkey ID or deletion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid password confirmation or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    PublishableKey:
      type: apiKey
      name: X-Authn-Publishable-Key
      in: header
      description: Client publishable key identifying application context
    SecretKey:
      type: apiKey
      name: X-Authn-Secret-Key
      in: header
      description: Backend secret key for tenant admin API calls
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Client user or console admin session JWT
  schemas:
...
    UserDTO:
      type: object
      description: Public user profile payload
      required:
      - id
      - email
      - email_verified
      - status
      - created_at
      properties:
        id:
          type: string
          example: usr_1a2b3c
        email:
          type: string
          format: email
          example: user@example.com
        email_verified:
          type: boolean
          example: true
        name:
          type: string
          example: Alex Smith
        status:
          type: string
          example: active
        created_at:
          type: string
          format: date-time
          example: 2026-08-01 12:00:00+00:00
    PolicyWarningDTO:
      type: object
      properties:
        requires_password_upgrade:
          type: boolean
        requires_email_verification:
          type: boolean
        missing_criteria:
          type: array
          items:
            type: string
    AuthResponse:
      type: object
      required:
      - user
      properties:
        user:
          $ref: '#/components/schemas/UserDTO'
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        policy_warning:
          $ref: '#/components/schemas/PolicyWarningDTO'
        mfa_required:
          type: boolean
        mfa_token:
          type: string
        methods:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
      - error
      - code
      properties:
        error:
          type: string
          example: invalid two-factor authentication code
        code:
          type: string
          example: invalid_mfa_code
...
    ConfirmTOTPResponse:
      type: object
      required:
      - message
      - recovery_codes_created
      properties:
        message:
          type: string
          example: TOTP 2FA enabled successfully
        recovery_codes:
          type: array
          items:
            type: string
          description: Freshly generated recovery codes (returned only upon initial setup/regeneration)
        recovery_codes_created:
          type: boolean
...
    BeginWebAuthnRegistrationResponse:
      type: object
      required:
      - options
      - session_id
      properties:
        options:
          type: object
          description: WebAuthn CredentialCreationOptions protocol object
          additionalProperties: true
        session_id:
          type: string
          description: Registration ceremony session ID
    WebAuthnRegisterFinishRequest:
      type: object
      properties:
        session_id:
          type: string
          description: Ceremony session ID
        name:
          type: string
          description: Human-readable passkey name label
    WebAuthnLoginBeginRequest:
      type: object
      required:
      - mfa_token
      properties:
        mfa_token:
          type: string
          description: MFA challenge token from login response
    BeginWebAuthnLoginResponse:
      type: object
      required:
      - options
      - session_id
      - user_id
      properties:
        options:
          type: object
          description: WebAuthn CredentialAssertionOptions protocol object
          additionalProperties: true
        session_id:
          type: string
          description: Assertion ceremony session ID
        user_id:
          type: string
          description: User ID being authenticated
    WebAuthnLoginFinishRequest:
      type: object
      properties:
        mfa_token:
          type: string
          description: MFA token
        session_id:
          type: string
          description: Assertion ceremony session ID
    PasskeyDTO:
      type: object
      required:
      - id
      - name
      - credential_id
      - sign_count
      - created_at
      properties:
        id:
          type: string
          example: mfa_passkey_1a2b3c
        name:
          type: string
          example: Work YubiKey
        credential_id:
          type: string
          example: dXNlck9wZW5JRA==
        sign_count:
          type: integer
          example: 42
        created_at:
          type: string
          format: date-time
          example: '2026-08-01T12:00:00Z'
        last_used_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-08-05T14:30:00Z'
        metadata:
          type: object
          description: Authenticator metadata (AAGUID, attestation type, transports, flags)
          additionalProperties: true
    ListPasskeysResponse:
      type: object
      required:
      - credentials
      properties:
        credentials:
          type: array
          items:
            $ref: '#/components/schemas/PasskeyDTO'
    PasskeyDeleteRequest:
      type: object
      properties:
        password:
          type: string
          description: Password step-up confirmation required if deleting the last remaining 2FA method
    DeletePasskeyResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          example: passkey successfully deleted
...